correcion en las pruebas unitarias, ya se pueden usar las constantes en las pruebas unitarias

+++ xcUtilities/src/xcVectorI2.cpp
@@ -11,26 +11,26 @@ namespace xcEngineSDK {
 
 
   VectorI2
-  VectorI2::operator - (const VectorI2& V) {
+  VectorI2::operator-(const VectorI2& V) {
 
     return VectorI2(m_x - V.m_x, m_y - V.m_y);
   }
 
   VectorI2
-  VectorI2::operator * (const VectorI2& V) {
+  VectorI2::operator*(const VectorI2& V) {
 
     return VectorI2(m_x * V.m_x, m_y * V.m_y);
   }
 
   VectorI2
-  VectorI2::operator = (const VectorI2& V) {
+  VectorI2::operator=(const VectorI2& V) {
     m_x = V.m_x;
     m_y = V.m_y;
     return *this;
   }
 
   bool
-  VectorI2::operator == (const VectorI2& V) {
+  VectorI2::operator==(const VectorI2& V) {
 
     if (m_x == V.m_x && m_y == V.m_y) {
       return true;
@@ -41,31 +41,31 @@ namespace xcEngineSDK {
   }
 
   VectorI2
-  VectorI2::operator + (const int32& V) {
+  VectorI2::operator+(const int32& V) {
 
     return VectorI2(m_x + V, m_y + V);
   }
 
   VectorI2
-  VectorI2::operator - (const int32& V) {
+  VectorI2::operator-(const int32& V) {
 
     return VectorI2(m_x - V, m_y - V);
   }
 
   VectorI2
-  VectorI2::operator * (const int32& V) {
+  VectorI2::operator*(const int32& V) {
 
     return VectorI2(m_x * V, m_y * V);
   }
 
   VectorI2
-  VectorI2::operator / (const int32& V) {
+  VectorI2::operator/(const int32& V) {
 
     return VectorI2(m_x / V, m_y / V);
   }
 
   VectorI2&
-  VectorI2::operator += (const VectorI2& V) {
+  VectorI2::operator +=(const VectorI2& V) {
 
     m_x += V.m_x;
     m_y += V.m_y;
@@ -111,7 +111,7 @@ namespace xcEngineSDK {
 
   int32
   VectorI2::magnitud() {
-    return static_cast <int32>(sqrt(m_x * m_x + m_y * m_y));
+    return (sqrt(m_x * m_x + m_y * m_y));
   }
 
   VectorI2
